+++ dorado/data_loader/DataLoader.cpp
@@ -360,7 +360,7 @@ int DataLoader::get_num_reads(std::string data_path,
         num_reads = std::min(num_reads, final_read_list.size());
     }
 
-    return num_reads;
+    return int(num_reads);
 }
 
 void DataLoader::load_read_channels(std::string data_path, bool recursive_file_loading) {
@@ -577,7 +577,7 @@ uint16_t DataLoader::get_sample_rate(std::string data_path, bool recursive_file_
             } else if (ext == ".fast5") {
                 H5Easy::File file(entry.path().string(), H5Easy::File::ReadOnly);
                 HighFive::Group reads = file.getGroup("/");
-                int num_reads = reads.getNumberObjects();
+                int num_reads = int(reads.getNumberObjects());
 
                 if (num_reads > 0) {
                     auto read_id = reads.getObjectName(0);
@@ -732,7 +732,7 @@ void DataLoader::load_pod5_reads_from_file(const std::string& path) {
         for (std::size_t row = 0; row < batch_row_count; ++row) {
             // TODO - check the read ID here, for each one, only send the row if it is in the list of ones we care about
 
-            if (can_process_pod5_row(batch, row, m_allowed_read_ids, m_ignored_read_ids)) {
+            if (can_process_pod5_row(batch, int(row), m_allowed_read_ids, m_ignored_read_ids)) {
                 futures.push_back(pool.push(process_pod5_read, row, batch, file, path,
                                             &m_reads_by_channel, &m_read_id_to_index));
             }
@@ -757,7 +757,7 @@ void DataLoader::load_fast5_reads_from_file(const std::string& path) {
     // Read the file into a vector of torch tensors
     H5Easy::File file(path, H5Easy::File::ReadOnly);
     HighFive::Group reads = file.getGroup("/");
-    int num_reads = reads.getNumberObjects();
+    int num_reads = int(reads.getNumberObjects());
 
     for (int i = 0; i < num_reads && m_loaded_read_count < m_max_reads; i++) {
         auto read_id = reads.getObjectName(i);
@@ -825,7 +825,7 @@ void DataLoader::load_fast5_reads_from_file(const std::string& path) {
                                                  static_cast<uint32_t>(start_time / sampling_rate));
 
         auto new_read = std::make_unique<SimplexRead>();
-        new_read->read_common.sample_rate = sampling_rate;
+        new_read->read_common.sample_rate = uint64_t(sampling_rate);
         new_read->read_common.raw_data = samples;
         new_read->digitisation = digitisation;
         new_read->range = range;
