+++ dorado/read_pipeline/BasecallerNode.cpp
@@ -115,7 +115,7 @@ void BasecallerNode::basecall_current_batch(int worker_id) {
     NVTX3_FUNC_RANGE();
     auto model_runner = m_model_runners[worker_id];
     dorado::stats::Timer timer;
-    auto decode_results = model_runner->call_chunks(m_batched_chunks[worker_id].size());
+    auto decode_results = model_runner->call_chunks(int(m_batched_chunks[worker_id].size()));
     m_call_chunks_ms += timer.GetElapsedMS();
 
     for (size_t i = 0; i < m_batched_chunks[worker_id].size(); i++) {
@@ -195,7 +195,7 @@ void BasecallerNode::basecall_worker_thread(int worker_id) {
     at::InferenceMode inference_mode_guard;
 
     auto last_chunk_reserve_time = std::chrono::system_clock::now();
-    int batch_size = m_model_runners[worker_id]->batch_size();
+    int batch_size = int(m_model_runners[worker_id]->batch_size());
     while (true) {
         std::unique_ptr<BasecallingChunk> chunk;
         const auto pop_status = m_chunks_in.try_pop_until(
@@ -341,7 +341,7 @@ void BasecallerNode::start_threads() {
     for (int i = 0; i < static_cast<int>(num_workers); i++) {
         m_basecall_workers[i] = std::thread([this, i] { basecall_worker_thread(i); });
     }
-    m_num_active_model_runners = num_workers;
+    m_num_active_model_runners = int(num_workers);
 }
 
 void BasecallerNode::terminate_impl() {
@@ -381,14 +381,14 @@ stats::NamedStats BasecallerNode::sample_stats() const {
         const auto runner_stats = stats::from_obj(*runner);
         stats.insert(runner_stats.begin(), runner_stats.end());
     }
-    stats["batches_called"] = m_num_batches_called;
-    stats["partial_batches_called"] = m_num_partial_batches_called;
-    stats["call_chunks_ms"] = m_call_chunks_ms;
-    stats["called_reads_pushed"] = m_called_reads_pushed;
-    stats["working_reads_items"] = m_working_reads_size;
-    stats["working_reads_signal_mb"] = m_working_reads_signal_bytes / (1024 * 1024);
-    stats["bases_processed"] = m_num_bases_processed;
-    stats["samples_processed"] = m_num_samples_processed;
+    stats["batches_called"] = double(m_num_batches_called);
+    stats["partial_batches_called"] = double(m_num_partial_batches_called);
+    stats["call_chunks_ms"] = double(m_call_chunks_ms);
+    stats["called_reads_pushed"] = double(m_called_reads_pushed);
+    stats["working_reads_items"] = double(m_working_reads_size);
+    stats["working_reads_signal_mb"] = double(m_working_reads_signal_bytes) / double((1024 * 1024));
+    stats["bases_processed"] = double(m_num_bases_processed);
+    stats["samples_processed"] = double(m_num_samples_processed);
     return stats;
 }
