+++ dorado/read_pipeline/HtsWriter.cpp
@@ -38,7 +38,7 @@ HtsWriter::HtsWriter(const std::string& filename, OutputMode mode, size_t thread
         throw std::runtime_error("Could not open file: " + filename);
     }
     if (m_file->format.compression == bgzf) {
-        auto res = bgzf_mt(m_file->fp.bgzf, threads, 128);
+        auto res = bgzf_mt(m_file->fp.bgzf, int(threads), 128);
         if (res < 0) {
             throw std::runtime_error("Could not enable multi threading for BAM generation.");
         }
@@ -161,7 +161,7 @@ int HtsWriter::set_and_write_header(const sam_hdr_t* const header) {
 
 stats::NamedStats HtsWriter::sample_stats() const {
     auto stats = stats::from_obj(m_work_queue);
-    stats["unique_simplex_reads_written"] = m_processed_read_ids.size();
+    stats["unique_simplex_reads_written"] = double(m_processed_read_ids.size());
     stats["duplex_reads_written"] = m_duplex_reads_written.load();
     stats["split_reads_written"] = m_split_reads_written.load();
     return stats;
